requirements.txt
blinker==1.4
certifi==2020.4.5.2
chardet==3.0.4
Django==2.2.12
django-mongoengine==0.4.1
django-mongoengine-forms==0.4.6
django-rest-framework==0.1.0
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.11.0
dnspython==1.16.0
greenlet==0.4.16
idna==2.9
mongoengine==0.20.0
pymongo==3.10.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
